# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shop API Documentation",
        "description": "A modern REST API built with Next.js",
        "version": "1.0.0"
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "Product": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "cd533345-f1f3-48c9-a62e-7dc2da50c8f8",
                        "description": "Product ID",
                        "uniqueItems": true
                    },
                    "title": {
                        "type": "string",
                        "example": "T-Shirt Teslo",
                        "description": "Product Title",
                        "uniqueItems": true
                    },
                    "price": {
                        "type": "number",
                        "example": 0,
                        "description": "Product price"
                    },
                    "description": {
                        "type": "string",
                        "example": "Anim reprehenderit nulla in anim mollit minim irure commodo.",
                        "description": "Product description",
                        "default": null
                    },
                    "slug": {
                        "type": "string",
                        "example": "t_shirt_teslo",
                        "description": "Product SLUG - for SEO",
                        "uniqueItems": true
                    },
                    "stock": {
                        "type": "number",
                        "example": 10,
                        "description": "Product stock",
                        "default": 0
                    },
                    "sizes": {
                        "example": [
                            "M",
                            "XL",
                            "XXL"
                        ],
                        "description": "Product sizes",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "gender": {
                        "type": "string",
                        "example": "women",
                        "description": "Product gender"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "id",
                    "title",
                    "price",
                    "description",
                    "slug",
                    "stock",
                    "sizes",
                    "gender",
                    "tags",
                    "images"
                ]
            },
            "ProductCreate": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Product title (unique)",
                        "nullable": false,
                        "minLength": 1
                    },
                    "price": {
                        "type": "number"
                    },
                    "description": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "stock": {
                        "type": "number"
                    },
                    "sizes": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "gender": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "title",
                    "price",
                    "description",
                    "slug",
                    "stock",
                    "sizes",
                    "gender",
                    "tags",
                    "images"
                ]
            },
            "ProductUpdate": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Product title (unique)",
                        "nullable": false,
                        "minLength": 1
                    },
                    "price": {
                        "type": "number"
                    },
                    "description": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "stock": {
                        "type": "number"
                    },
                    "sizes": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "gender": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "UserCreate": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "description": "A valid email (unique)",
                        "nullable": false
                    },
                    "password": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "password",
                    "fullName"
                ]
            },
            "UserLogin": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "description": "A valid email",
                        "nullable": false
                    },
                    "password": {
                        "type": "string",
                        "description": "The password for this user",
                        "nullable": false
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            }
        }
    },
    "paths": {
        "/api/products": {
            "post": {
                "summary": "Create new product",
                "tags": [
                    "Products"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product was created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing data"
                    },
                    "403": {
                        "description": "Forbidden. Token related."
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "get": {
                "summary": "List products",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "name": "limit",
                        "required": true,
                        "in": "query",
                        "description": "How many rows do you need",
                        "schema": {
                            "default": 10,
                            "type": "number"
                        }
                    },
                    {
                        "name": "offset",
                        "required": true,
                        "in": "query",
                        "description": "How many rows do you want to skip",
                        "schema": {
                            "default": 0,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product list"
                    }
                }
            }
        },
        "/api/products/{idOrSlug}": {
            "get": {
                "summary": "Find product by ID or slug",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "name": "idOrSlug",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product found"
                    }
                }
            }
        },
        "/api/products/all/{term}": {
            "get": {
                "summary": "Search products",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "name": "term",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "patch": {
                "summary": "Update product",
                "tags": [
                    "Products"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "summary": "Delete product",
                "tags": [
                    "Products"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "summary": "Register",
                "tags": [
                    "Auth"
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registration successful"
                    },
                    "400": {
                        "description": "Missing required fields"
                    },
                    "409": {
                        "description": "User already exists"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "summary": "Login",
                "tags": [
                    "Auth"
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLogin"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login successful"
                    },
                    "400": {
                        "description": "Missing email or password"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/auth/check-status": {
            "get": {
                "summary": "Check authentication status",
                "tags": [
                    "Auth"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Check successful"
                    },
                    "401": {
                        "description": "Unauthorized / Invalid token"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/files/product/{imageName}": {
            "get": {
                "summary": "Get product image",
                "tags": [
                    "Files - Get and Upload"
                ],
                "parameters": [
                    {
                        "name": "imageName",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Image not found"
                    }
                }
            }
        },
        "/api/files/product": {
            "post": {
                "summary": "Upload product image",
                "tags": [
                    "Files - Get and Upload"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [],
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    }
}